Marker note is no longer needed! Added basepath settings option to make it easier to navigate

+++ BeatMerge/Items/Event.cs
@@ -3,7 +3,7 @@ namespace BeatMerge.Items
 {
     public class Event : ItemBase
     {
-        public int _type { get; set; }
-        public int _value { get; set; }
+        public double _type { get; set; }
+        public double _value { get; set; }
     }
 }
