+++ BeatMerge/Items/Obstacle.cs
@@ -4,9 +4,9 @@ namespace BeatMerge.Items
 {
     public class Obstacle : ItemBase
     {
-        public int _lineIndex { get; set; }
-        public int _type { get; set; }
-        public int _duration { get; set; }
-        public int _width { get; set; }
+        public double _lineIndex { get; set; }
+        public double _type { get; set; }
+        public double _duration { get; set; }
+        public double _width { get; set; }
     }
 }
